Try to Compare RFR and SVR Algorithm
In this notebook, i was try to compare two algorithm in machine learning which is Random Forest Regressor (RFR) and Support Vector Regressor (SVR). you know that Random Forest and SVM is a part of classification algorithm in machine learning, however i very curious to use it in Regression task. Please see how the results are...